Function that compares two strings

+++ 0x06-pointers_arrays_strings/3-strcmp.c
@@ -17,7 +17,7 @@ int _strcmp(char *s1, char *s2)
 	if (*s1 == '\0' && *s2 == '\0')
 		return (0);
 	else if (*s1 < *s2)
-		return (*s2 - *s1);
+		return ((int)*s2 - *s1);
 	else
-		return (*s2 - *s1);
+		return ((int)*s2 - *s1);
 }
